engine: split xmile.rs into focused submodules
The 4,799-line xmile.rs handled all XMILE format operations in one
file. Split it into an xmile/ directory with 5 focused modules:

- mod.rs (839 lines): shared types, XML helpers, top-level file I/O
- dimensions.rs (349 lines): graphical functions and dimension parsing
- model.rs (517 lines): Model struct, references, connections
- variables.rs (968 lines): variable types and datamodel conversion
- views.rs (2,215 lines): view elements, layout, and serialization

Each module handles a distinct aspect of the XMILE format. The views
module is the largest due to the inherent complexity of view element
layout and serialization, but it focuses on a single cohesive concern.
